igb: Starting point for an emulated Intel 82576 Gigabit Ethernet adapter w/SR/IOV support
A fresh start at implementing an emulation of the
Intel 82576 Gigabit Ethernet Adapter (igb/igbvf).

The previous attempt, see:

https://lists.gnu.org/archive/html/qemu-devel/2014-08/msg05110.html

used the e1000.c code in QEMU as basis.

Since then a more advanced implementation of the (newer) E1000e,
the Intel 82574 Gigabit Ethernet Controller, has been merged,
see:

https://lists.gnu.org/archive/html/qemu-devel/2016-02/msg04073.html

The E1000e hardware seems more similar to the IGB.
In addition the code is structured in a way that lends itself better
to sharing code, and the implementation is also more efficient.

This patch is based on a freshly rebased
version of my SR/IOV patch set.

On any recent Linux kernel, the igb driver will detect the emulated
device and get as far as to enable interrupts, which does not work
as the implementation is only partial.

I have been using a generic Qemu root port to test the device with,
eg. start with something like

   qemu-kvm -M q35 \
      ...
      -device pcie-root-port,slot=2,id=pcie_port.2 \
      -device igb,bus=pcie_port.2

It is then possible to enable and disable VFs by either

1) loading/unloading the igb driver and use a nonzero value
   for the module parameter, eg

   rmmod igb
   modprobe igb max_vfs=2

2) or by using PCIe config space commands via /sys, eg.

   echo 7 > /sys/bus/pci/devices/0000\:03\:00.0/sriov_numvfs
   echo 0 > /sys/bus/pci/devices/0000\:03\:00.0/sriov_numvfs

It is still far from a working network device in any way,
but should hopefully be an improved starting point for a
complete implementation, and useful as a more up-to-date
device to demonstrate/play with SR/IOV on a VM.

Signed-off-by: Knut Omang <knuto@ifi.uio.no>
[Fix meson build]
Signed-off-by: Marcel Apfelbaum <marcel@redhat.com>